refactor(seed): addressed comments and refactored file structure
